:hammer: Refactor: Parser 클래스는 파싱만 하도록 Regex 생성 로직 분리
